Handling ScriptureDocuments correctly
