+++ packages/punctuation-checker/test/test-utils.ts
@@ -1,12 +1,24 @@
-import { DocumentManager, Position, Range, TextDocument } from '@sillsdev/lynx';
+import { DocumentManager, Position, Range, ScriptureDocument, TextDocument } from '@sillsdev/lynx';
+import { UsfmDocumentFactory } from '@sillsdev/lynx-usfm';
 
 // Just returns a document with the text of the URI
 // (allows for much more concise testing, since we avoid creating
 // a document and then referring to when we ask for Diagnostics)
-export class StubDocumentManager extends DocumentManager<TextDocument> {
+export class StubTextDocumentManager extends DocumentManager<TextDocument> {
   get(text: string): Promise<TextDocument | undefined> {
     return new Promise<TextDocument>((resolve) => {
       resolve(new StubSingleLineTextDocument(text));
+    });
+  }
+}
+export class StubScriptureDocumentManager extends DocumentManager<ScriptureDocument> {
+  constructor(private readonly usfmDocumentFactory: UsfmDocumentFactory) {
+    super(usfmDocumentFactory);
+  }
+
+  get(text: string): Promise<ScriptureDocument | undefined> {
+    return new Promise<ScriptureDocument>((resolve) => {
+      resolve(this.usfmDocumentFactory.create('test', 'usfm', 1, text));
     });
   }
 }
